+++ kivakit-interfaces/src/main/java/com/telenav/kivakit/interfaces/collection/Sequence.java
@@ -91,8 +91,9 @@ import java.util.function.Function;
  * @see Indexable
  * @see Matcher
  */
+@SuppressWarnings("unused")
 @UmlClassDiagram(diagram = DiagramCollection.class)
-public interface Sequence<Element> extends Iterable<Element>
+public interface Sequence<Element> extends Iterable<Element>, Joinable<Element>
 {
     /**
      * @return True if all elements in this sequence match the given matcher
@@ -168,24 +169,18 @@ public interface Sequence<Element> extends Iterable<Element>
     default List<Element> asList()
     {
         var list = new ArrayList<Element>();
-        for (var element : this)
-        {
-            list.add(element);
-        }
+        forEach(list::add);
         return list;
     }
 
     /**
-     * @return This sequence as a list
+     * @return This sequence as a set
      */
     default Set<Element> asSet()
     {
-        var set = new HashSet<Element>();
-        for (var element : this)
-        {
-            set.add(element);
-        }
-        return set;
+        var list = new HashSet<Element>();
+        forEach(list::add);
+        return list;
     }
 
     default Element find(Class<? extends Element> type)
@@ -288,53 +283,6 @@ public interface Sequence<Element> extends Iterable<Element>
     default Iterator<Element> iterator()
     {
         return asIterator();
-    }
-
-    /**
-     * @return The elements in this sequence joined as a string with the given separator
-     */
-    default String join(char separator)
-    {
-        return join(String.valueOf(separator));
-    }
-
-    /**
-     * @return The elements in this sequence joined as a string with the given separator
-     */
-    default String join(String separator)
-    {
-        return join(separator, Object::toString);
-    }
-
-    /**
-     * @return The elements in this sequence joined as a string with the given separator or the default value if this
-     * sequence is empty
-     */
-    default String join(String separator, String defaultValue)
-    {
-        if (first() == null)
-        {
-            return defaultValue;
-        }
-        return join(separator, Object::toString);
-    }
-
-    /**
-     * @return The elements in this sequence transformed into strings by the given function and joined together with the
-     * given separator
-     */
-    default String join(String separator, Function<Element, String> toString)
-    {
-        var builder = new StringBuilder();
-        for (var value : this)
-        {
-            if (builder.length() > 0)
-            {
-                builder.append(separator);
-            }
-            builder.append(toString.apply(value));
-        }
-        return builder.toString();
     }
 
     /**
