+++ kivakit-network/http/src/main/java/com/telenav/kivakit/network/http/HttpMethod.java
@@ -1,0 +1,41 @@
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+//
+// © 2011-2021 Telenav, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+package com.telenav.kivakit.network.http;
+
+/**
+ * @author jonathanl (shibo)
+ */
+public enum HttpMethod
+{
+    GET,
+    POST,
+    DELETE,
+    HEAD,
+    PUT,
+    CONNECT,
+    OPTIONS,
+    TRACE,
+    PATCH,
+    NONE;
+
+    public static HttpMethod parse(String text)
+    {
+        return valueOf(text.toUpperCase());
+    }
+}
